requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
channels==4.0.0
charset-normalizer==3.2.0
Django==4.2.5
django-admin-rangefilter==0.11.1
django-bootstrap-v5==1.0.11
django-filter==23.2
django-jazzmin==2.6.0
django-nextjs==2.4.0
django-rangefilter==0.9.1
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.4
multidict==6.0.4
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
